# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Integration Swagger in Larave",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "vyacheslavkarmazin@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Return user",
                "description": "Get users",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::getAll",
                "responses": {
                    "200": {
                        "description": "Array of user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "user id",
                                                "type": "integer",
                                                "nullable": true
                                            },
                                            "name": {
                                                "description": "Name of user",
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "email": {
                                                "description": "Email",
                                                "type": "string",
                                                "example": "test@gmail.com",
                                                "nullable": false
                                            },
                                            "password": {
                                                "description": "Password",
                                                "type": "string",
                                                "nullable": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Return user",
                "description": "Create user",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of user",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "example": "test@gmail.com",
                                        "nullable": false
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "nullable": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Array of user"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Return user",
                "description": "Get user",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::getById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "user id",
                                            "type": "integer",
                                            "nullable": true
                                        },
                                        "name": {
                                            "description": "Name of user",
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "email": {
                                            "description": "Email",
                                            "type": "string",
                                            "example": "test@gmail.com",
                                            "nullable": false
                                        },
                                        "password": {
                                            "description": "Password",
                                            "type": "string",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Return user",
                "description": "Create user",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "user id",
                                            "type": "integer",
                                            "nullable": true
                                        },
                                        "name": {
                                            "description": "Name of user",
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "email": {
                                            "description": "Email",
                                            "type": "string",
                                            "example": "test@gmail.com",
                                            "nullable": false
                                        },
                                        "password": {
                                            "description": "Password",
                                            "type": "string",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Value: Bearer {jwt}",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}